chore: remove compiled caches
